backend/requirements.txt
Flask
SQLAlchemy
PyMySQL
python-dotenv
alembic
cryptography
